End of Part 2| CQRS| Product|Category|Validator|Exception|Response

+++ src/Core/Ecommerce.Application/Features/Categories/Handlers/Command/DeleteCategoryCommandHandler.cs
@@ -1,10 +1,12 @@
-﻿using Ecommerce.Application.Features.Categories.Requests.Command;
+﻿using Ecommerce.Application.Exceptions;
+using Ecommerce.Application.Features.Categories.Requests.Command;
 using Ecommerce.Application.Persistance.Contracts;
+using Ecommerce.Domain.Entities;
 using MediatR;
 
 namespace Ecommerce.Application.Features.Categories.Handlers.Command
 {
-    public class DeleteCategoryCommandHandler : IRequestHandler<DeleteCategoryCommand>
+    public class DeleteCategoryCommandHandler : IRequestHandler<DeleteCategoryCommand,Unit>
     {
         private readonly ICategoryRepository _repository;
         private readonly IMapper _mapper;
@@ -15,14 +17,20 @@ namespace Ecommerce.Application.Features.Categories.Handlers.Command
             _mapper = mapper;
         }
 
-        public async Task Handle(DeleteCategoryCommand request, CancellationToken cancellationToken)
+        public async Task<Unit> Handle(DeleteCategoryCommand request, CancellationToken cancellationToken)
         {
             // get category by id
             var oldCategory = await _repository.GetAsync(request.Id);
+            if(oldCategory is null)
+            {
+                throw new NotFoundException(nameof(Category), request.Id);
+            }
+
 
             // remove
             await _repository.DeleteAsync(oldCategory.Id);
 
+            return Unit.Value;
         }
     }
 }
